Split files, create test
