+++ src/main/java/ch/njol/skript/expressions/arithmetic/NumberExpressionInfo.java
@@ -1,0 +1,39 @@
+/**
+ *   This file is part of Skript.
+ *
+ *  Skript is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  Skript is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with Skript.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * Copyright Peter Güttinger, SkriptLang team and contributors
+ */
+package ch.njol.skript.expressions.arithmetic;
+
+import org.bukkit.event.Event;
+
+import ch.njol.skript.lang.Expression;
+
+public class NumberExpressionInfo implements ArithmeticGettable {
+	
+	private final Expression<? extends Number> expression;
+	
+	public NumberExpressionInfo(Expression<? extends Number> expression) {
+		this.expression = expression;
+	}
+	
+	@Override
+	public Number get(Event event, boolean integer) {
+		Number number = expression.getSingle(event);
+		return number != null ? number : 0;
+	}
+	
+}
