+++ src/test/skript/tests/regressions/535-math is done in the wrong order.sk
@@ -1,0 +1,10 @@
+test "math order":
+	assert 1 + 1 = 2 with "basic math ##1 failed"
+	assert 5 - 3 = 2 with "basic math ##2 failed"
+
+	assert 5 - 3 - 1 = 1 with "basic chained math ##1 failed"
+	assert 5-3-2 = 0 with "basic chained math ##2 failed"
+	assert 10 - 1 - 5 = 4 with "basic chained math ##3 failed"
+
+	assert (9 - 1) - 3 = 5 with "basic chained math with parentheses ##1 failed"
+	assert 9 - (1 - 3) = 11 with "basic chained math with parentheses ##2 failed"
